Export both ReactMounter and ReactView
