+++ src/hoc/ReactMounter/index.js
@@ -1,0 +1,2 @@
+import ReactMounter from './ReactMounter'
+export {ReactMounter}
