+++ src/hoc/ReactMounter/util/__tests__/buildComponents.test.js
