+++ src/hoc/ReactMounter/util/__tests__/buildEnhancedComponent.test.js
