+++ src/hoc/ReactMounter/util/buildComponents.js
