+++ src/hoc/ReactMounter/util/buildEnhancedComponent.js
