+++ src/hoc/ReactView/index.js
@@ -1,0 +1,2 @@
+import ReactView from './ReactView'
+export {ReactView}
