+++ src/hoc/index.js
@@ -1,0 +1,2 @@
+export {ReactMounter} from './ReactMounter'
+export {ReactView} from './ReactView'
