--- src/mixins/ReactMixin/index.js
@@ -1,2 +1,0 @@
-import ReactMixin from './ReactMixin'
-export {ReactMixin}
