--- src/mixins/ReactView/index.js
@@ -1,2 +1,0 @@
-import ReactView from './ReactView'
-export default ReactView
