--- src/mixins/index.js
@@ -1,1 +1,0 @@
-export {ReactMixin} from './ReactMixin'
