+++ stories/ReactMounter/templates/template.js
